Moved export button position to source loader
