+++ annotation/view/gui/Controls.py
@@ -306,69 +306,17 @@ class SequenceClassificationControls:
             self.controller.set_correct_classifications(self.classification_selector.value)
 
 
-class ExportControls:
-    FILENAME: str = "sequence_annotation"
-
-    def __init__(self, controller: AnnotationController, reset_visibility_fn: Callable):
-        self.controller: AnnotationController = controller
-        self.reset_visibility_fn = reset_visibility_fn
-
-        self.show_options_button = Button(name="Show export options", button_type="success", button_style="outline")
-        file_download_buttons = [FileDownload(
-            label=f"Export to {ftype}",
-            filename=f"{ExportControls.FILENAME}.{ftype}",
-            callback=bind(self.export, filetype=ftype),
-            button_type="primary",
-            button_style="outline") for ftype in self.controller.get_export_file_formats()]
-        self.export_buttons = Row(
-            *file_download_buttons,
-            visible=False
-        )
-
-        self.show_options_button.on_click(self.toggle_options_visibility)
-
-        self.component = Column(
-            self.show_options_button,
-            self.export_buttons,
-            align="start"
-        )
-
-    def get_component(self):
-        return self.component
-
-    def set_visibility(self, is_visible: bool):
-        self.component.visible = is_visible
-
-    def toggle_visibility(self):
-        self.component.visible = not self.component.visible
-
-    def export(self, filetype: str, *_):
-        return self.controller.export(filetype)
-
-    def toggle_options_visibility(self, *_):
-        self.export_buttons.visible = not self.export_buttons.visible
-        if self.export_buttons.visible:
-            self.show_options_button.name = "Hide export options"
-            self.show_options_button.button_style = "solid"
-        else:
-            self.show_options_button.name = "Show export options"
-            self.show_options_button.button_style = "outline"
-
-
 class Controls:
     def __init__(self, controller: AnnotationController):
         self.controller: AnnotationController = controller
         self.clause_sequence_controls = ClauseSequenceControls(self.controller, self.add_sequence_controls)
         self.sequence_classification_controls = SequenceClassificationControls(self.controller)
         self.add_sequence_controls = AddSequenceControls(self.controller, self.reset_visibility)
-        self.export_controls = ExportControls(self.controller, self.reset_visibility)
 
         self.component = Column(
             self.clause_sequence_controls.get_component(),
             self.sequence_classification_controls.get_component(),
             self.add_sequence_controls.get_component(),
-            Divider(),
-            self.export_controls.get_component(),
             styles=controls_style,
             sizing_mode="stretch_width"
         )
@@ -396,6 +344,5 @@ class Controls:
         self.add_sequence_controls.set_visibility(False)
         self.clause_sequence_controls.set_visibility(True)
         self.sequence_classification_controls.set_visibility(True)
-        self.export_controls.set_visibility(True)
 
         self.update_display()
